Properly fixed parsing HEAD responses.
My first attempt added a bug in the parsing of normal GET responses
where, if the timing was correct they would be considered HEAD
responses and the body lost.
This version requires the user to tell fast-http when we are
expecting a HEAD response as there is no standards complient way to
tell a GET response from a HEAD response.